requirements.txt
cryptography==2.6.1
python-dotenv==0.10.1
pyinstaller==3.4
pytest==4.3.1
pytest-mock==1.10.1
gnureadline==6.3.8
PyNaCl==1.3.0